Revamped program to a more simple version
